Added CustomthemeData, Integration testing, UI details

+++ Movie App Test/movie_app/lib/models/movie.dart
@@ -38,7 +38,7 @@ class Movie {
     originalLanguage = json['original_language'];
     originalTitle = json['original_title'];
     overview = json['overview'];
-    popularity = json['popularity'];
+    popularity = json['popularity'].toDouble();
     posterPath = json['poster_path'];
     releaseDate = json['release_date'];
     title = json['title'];
